Removed CustomViewIdProvider from SafeIdGenerator
