+++ Assets/Countly/Plugins/CountlySDK/Helpers/SafeIDGenerator.cs
@@ -11,16 +11,4 @@ namespace Plugins.CountlySDK
             return CountlyUtils.SafeRandomVal();
         }
     }
-    /// <summary>
-    /// CustomViewIdProvider class is for testing purposes
-    /// </summary>
-    public class CustomViewIdProvider : ISafeIDGenerator
-    {
-        private int viewCount;
-        public string GenerateValue()
-        {
-            viewCount++;
-            return "idv" + viewCount.ToString();
-        }
-    }
 }
