+++ Assets/Tests/PlayModeTests/CustomViewIdProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d39d15759e29a4996b24a6a8f04f1120
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
